Fixed issue where DS_Store was making folder considered non-empty

+++ Editor/Core/Components/AssetSwitcher/AssetSwitcherOperations.cs
@@ -91,6 +91,12 @@ namespace Miniclip.ShapeShifter.Switcher
         public static void CopyFromSkinsToUnity(DirectoryInfo directory)
         {
             string guid = directory.Name;
+
+            if (!FileUtils.DoesFolderExistAndHaveFiles(directory.FullName))
+            {
+                FileUtils.SafeDelete(directory.FullName);
+                return;
+            }
             
             AssetSwitcher.DeleteAssetInternalCopy(guid);
